MET-4284 Improve syncrhonization report in sandbox(#60)
    MET-4284 change data type to support null in total amount of records

    MET-4284 change order of condition and fix sonar reported error on it

    MET-4284 change data type to support null as total quantity records when harvesting identifiers due is unknown value

    MET-4284 rename constants name to a friendly description

    MET-4284 add unit test coverage

    MET-4284 add message to publish portal when harvesting identifiers

    MET-4284 add total records for normal harvest

    MET-4284 total records for harvest oai pmh

+++ src/main/java/eu/europeana/metis/sandbox/repository/DatasetRepository.java
@@ -28,7 +28,7 @@ public interface DatasetRepository extends JpaRepository<DatasetEntity, Integer>
    */
   @Modifying
   @Query("UPDATE DatasetEntity dataset SET dataset.recordsQuantity = ?2 WHERE dataset.datasetId = ?1")
-  void updateRecordsQuantity(int datasetId, int quantity);
+  void updateRecordsQuantity(int datasetId, Long quantity);
 
   /**
    * Sets to true the boolean recordLimitExceeded
